+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.0.0
+
+  - Sencha Touch 2.4.1 GPL
+  - Assets pipeline full support
+
+## 1.2.0
+  
+  - Grand refactoring initial
+
 ## 1.0.1 (17 Feb 2012)
 
   - Sass